+++ sc-kpm/ui/ui.cpp
@@ -20,7 +20,7 @@ const sc_version required_version = {0, 1, 0, ""};
 sc_memory_context * s_default_ctx = 0;
 
 // ------------------- Module ------------------------------
-_SC_EXT_EXTERN sc_result initialize()
+_SC_EXT_EXTERN sc_result sc_module_initialize()
 {
   s_default_ctx = sc_memory_context_new(sc_access_lvl_make_min);
 
@@ -36,7 +36,7 @@ _SC_EXT_EXTERN sc_result initialize()
   return SC_RESULT_OK;
 }
 
-_SC_EXT_EXTERN sc_result shutdown()
+_SC_EXT_EXTERN sc_result sc_module_shutdown()
 {
 
   ui_shutdown_translators();